run-command.c: don't inherit pipe handles in children
When Git runs a long-running filter process (such as git-lfs), it opens
pipe handles to communicate, and Git will remain running until those
handles are closed. The intention is that the filter process will close
the handles when it terminates, and Git will stop immediately afterward.

However, if a post-checkout or post-commit hook starts a background
process (such as one to update a ctags index), then the background
process will inherit the filter process's pipe handles and keep them
open. Thus, Git will wait not only for the filter process to terminate,
but also for the background hook to terminate (which means it's not
exactly a "background" process anymore). The visible symptom of this
issue is that takes a while for the command prompt to come back after a
git command.

The fix is to mark the unneeded pipe handles with close-on-exec so that
subsequent processes don't inherit them.

Those affected by this issue can work around it in older Git versions by
modifying the hook script to close any unneeded handles, such as like
this:

    for fd in $(ls /proc/$$/fd); do
      case "$fd" in
        0|1|2|255)
          ;;
        *)
          eval "exec $fd>&-"
          ;;
      esac
    done